fix: address Aristotel review — proper layering, file size compliance
