+++ bridge/app/lib/src/bridge/repositories/session_repository.dart
@@ -104,6 +104,19 @@ class SessionRepository {
         .toList(growable: false);
   }
 
+  Future<bool> hasOtherActiveSessionsSharing({
+    required String sessionId,
+    required String? worktreePath,
+    required String? branchName,
+  }) async {
+    final sessions = await _sessionDao.getOtherActiveSessionsSharing(
+      sessionId: sessionId,
+      worktreePath: worktreePath,
+      branchName: branchName,
+    );
+    return sessions.isNotEmpty;
+  }
+
   Future<String?> getProjectPath({required String projectId}) async {
     try {
       final project = await _plugin.getProject(projectId);
