+++ bridge/app/lib/src/bridge/routing/update_session_archive_status_handler.dart
@@ -8,6 +8,7 @@ import "package:sesori_shared/sesori_shared.dart";
 import "../persistence/daos/session_dao.dart";
 import "../persistence/tables/session_table.dart";
 import "../repositories/mappers/plugin_session_mapper.dart";
+import "../repositories/session_repository.dart";
 import "../worktree_service.dart";
 import "request_handler.dart";
 import "worktree_cleanup.dart";
@@ -17,14 +18,17 @@ class UpdateSessionArchiveStatusHandler extends BodyRequestHandler<UpdateSession
   final BridgePlugin _plugin;
   final WorktreeService _worktreeService;
   final SessionDao _sessionDao;
+  final SessionRepository _sessionRepository;
 
   UpdateSessionArchiveStatusHandler({
     required BridgePlugin plugin,
     required WorktreeService worktreeService,
     required SessionDao sessionDao,
+    required SessionRepository sessionRepository,
   }) : _plugin = plugin,
        _worktreeService = worktreeService,
        _sessionDao = sessionDao,
+       _sessionRepository = sessionRepository,
        super(
          HttpMethod.patch,
          "/session/update/archive",
@@ -146,6 +150,7 @@ class UpdateSessionArchiveStatusHandler extends BodyRequestHandler<UpdateSession
       )) {
         final cleanupResult = await performWorktreeCleanup(
           worktreeService: _worktreeService,
+          sessionRepository: _sessionRepository,
           sessionId: sessionDto.sessionId,
           projectId: projectId,
           worktreePath: worktreePath,
