+++ bridge/app/test/bridge/routing/routing_test_helpers.dart
@@ -526,6 +526,12 @@ class _NoopSessionRepository implements SessionRepository {
   Future<List<StoredSession>> getStoredSessionsByProjectId({required String projectId}) async =>
       const <StoredSession>[];
   @override
+  Future<bool> hasOtherActiveSessionsSharing({
+    required String sessionId,
+    required String? worktreePath,
+    required String? branchName,
+  }) async => false;
+  @override
   Future<String?> getProjectPath({required String projectId}) async => null;
 }
 
@@ -614,6 +620,13 @@ class FakeSessionRepository implements SessionRepository {
     final sessions = await _sessionDao.getSessionsByProject(projectId: projectId);
     return sessions.map((s) => StoredSession(id: s.sessionId, branchName: s.branchName)).toList(growable: false);
   }
+
+  @override
+  Future<bool> hasOtherActiveSessionsSharing({
+    required String sessionId,
+    required String? worktreePath,
+    required String? branchName,
+  }) async => false;
 
   @override
   Future<String?> getProjectPath({required String projectId}) async {
