+++ bridge/app/test/bridge/services/pr_sync_service_test.dart
@@ -414,5 +414,12 @@ class _FakeSessionRepository implements SessionRepository {
   }
 
   @override
+  Future<bool> hasOtherActiveSessionsSharing({
+    required String sessionId,
+    required String? worktreePath,
+    required String? branchName,
+  }) async => false;
+
+  @override
   Future<String?> getProjectPath({required String projectId}) async => null;
 }
